Split extraction methods and ensure they return null when no match is found.
Open Graph and selector-based extraction are now separate forms of extraction
instead of a single, "fallback" extraction method.
